Fix incorrect parameter type

+++ exercises/getting-started/shaders/exercise1/src/shader.h
@@ -17,7 +17,7 @@ public:
 	// Uniform setters.
 	void setBool(const std::string &name, bool value) const;
 	void setInt(const std::string &name, int value) const;
-	void setFloat(const std::string &name, int value) const;
+	void setFloat(const std::string &name, float value) const;
 
 private:
 	unsigned int program;
