+++ projects/getting-started/shaders/src/shader.cpp
@@ -48,7 +48,7 @@ void Shader::setInt(const std::string &name, int value) const {
 	glUniform1i(glGetUniformLocation(program, name.c_str()), value);
 }
 
-void Shader::setFloat(const std::string &name, int value) const {
+void Shader::setFloat(const std::string &name, float value) const {
 	glUniform1f(glGetUniformLocation(program, name.c_str()), value);
 }
